+++ lib/model/event2.dart
@@ -527,11 +527,15 @@ class Event2SurveyDetails {
 /// Event2Grouping
 
 class Event2Grouping {
+  static final String _ignoreFieldValue = 'ignore';
+  static final String _noneFieldValue = 'none';
+
   final Event2GroupingType? type;
   final String? superEventId;
   final String? recurrenceId;
+  final dynamic displayAsIndividual;
 
-  Event2Grouping({this.type, this.superEventId, this.recurrenceId});
+  Event2Grouping({this.type, this.superEventId, this.recurrenceId, this.displayAsIndividual});
 
   factory Event2Grouping.superEvent(String? id) => Event2Grouping(
     type: Event2GroupingType.superEvent,
@@ -543,29 +547,59 @@ class Event2Grouping {
     recurrenceId: id,
   );
 
+  static List<Event2Grouping> superEvents() => <Event2Grouping>[
+    Event2Grouping(type: Event2GroupingType.superEvent, superEventId: _noneFieldValue, recurrenceId: _ignoreFieldValue, displayAsIndividual: _ignoreFieldValue)
+  ];
+
+  static List<Event2Grouping> individualEvents() => <Event2Grouping>[
+    // Events without grouping type
+    Event2Grouping(type: Event2GroupingType.none, superEventId: _ignoreFieldValue, recurrenceId: _ignoreFieldValue, displayAsIndividual: _ignoreFieldValue),
+    // Main Super Events
+    Event2Grouping(type: Event2GroupingType.superEvent, superEventId: _noneFieldValue, recurrenceId: _ignoreFieldValue, displayAsIndividual: _ignoreFieldValue),
+    // Sub Events when display as individual is true
+    Event2Grouping(type: Event2GroupingType.superEvent, superEventId: _ignoreFieldValue, recurrenceId: _ignoreFieldValue, displayAsIndividual: 'true'), //explicitly String
+    // Main Recurring events
+    Event2Grouping(type: Event2GroupingType.recurrence, superEventId: _ignoreFieldValue, recurrenceId: _noneFieldValue, displayAsIndividual: _ignoreFieldValue)
+  ];
+
   static Event2Grouping? fromJson(Map<String, dynamic>? json) =>
     (json != null) ? Event2Grouping(
       type: event2GroupingTypeFromString(JsonUtils.stringValue(json['grouping_type'])) ,
       superEventId: JsonUtils.stringValue(json['super_event_id']),
       recurrenceId: JsonUtils.stringValue(json['group_id']),
+      displayAsIndividual: JsonUtils.boolValue(json['display_as_individual']),
     ) : null;
 
   Map<String, dynamic> toJson() => {
     'grouping_type': event2GroupingTypeToString(type),
     'super_event_id': superEventId,
     'group_id': recurrenceId,
+    'display_as_individual': (displayAsIndividual is bool) ? displayAsIndividual : displayAsIndividual?.toString(),
   };
+
+  static List<dynamic>? listToJson(List<Event2Grouping>? contentList) {
+    List<dynamic>? jsonList;
+    if (contentList != null) {
+      jsonList = <dynamic>[];
+      for (dynamic contentEntry in contentList) {
+        jsonList.add(contentEntry?.toJson());
+      }
+    }
+    return jsonList;
+  }
 
   @override
   bool operator==(dynamic other) =>
     (other is Event2Grouping) &&
     (type == other.type) &&
+    (displayAsIndividual == other.displayAsIndividual) &&
     (superEventId == other.superEventId) &&
     (recurrenceId == other.recurrenceId);
 
   @override
   int get hashCode =>
     (type?.hashCode ?? 0) ^
+    (displayAsIndividual?.hashCode ?? 0) ^
     (superEventId?.hashCode ?? 0) ^
     (recurrenceId?.hashCode ?? 0);
 }
@@ -1176,7 +1210,7 @@ String? event2SortOrderToOption(Event2SortOrder? value) {
 ///////////////////////////////
 /// Event2GroupingType
 
-enum Event2GroupingType { superEvent, recurrence }
+enum Event2GroupingType { superEvent, recurrence, none }
 
 Event2GroupingType? event2GroupingTypeFromString(String? value) {
   if (value == 'super_events') {
@@ -1184,6 +1218,9 @@ Event2GroupingType? event2GroupingTypeFromString(String? value) {
   }
   else if (value == 'repeatable') {
     return Event2GroupingType.recurrence;
+  }
+  else if (value == 'none') {
+    return Event2GroupingType.none;
   }
   else {
     return null;
@@ -1194,6 +1231,7 @@ String? event2GroupingTypeToString(Event2GroupingType? value) {
   switch (value) {
     case Event2GroupingType.superEvent: return 'super_events';
     case Event2GroupingType.recurrence: return 'repeatable';
+    case Event2GroupingType.none: return 'none';
     default: return null;
   }
 }